+++ packages/integration-sdk-cli/src/commands/index.ts
@@ -4,3 +4,4 @@ export * from './sync';
 export * from './run';
 export * from './document';
 export * from './types';
+export * from './visualize-types';