+++ packages/integration-sdk-cli/src/index.ts
@@ -1,6 +1,6 @@
 import { createCommand } from 'commander';
 
-import { collect, visualize, sync, run, document, types } from './commands';
+import { collect, visualize, sync, run, document, types, visualizeTypes } from './commands';
 
 export function createCli() {
   return createCommand()
@@ -9,5 +9,6 @@ export function createCli() {
     .addCommand(sync())
     .addCommand(run())
     .addCommand(types())
+    .addCommand(visualizeTypes())
     .addCommand(document());
 }